Improve path handling
Installation of TurtleBlocks as a standalone app is currently broken,
and this is the first step towards fixing that.

Application resources are divided into "share" and "lib" parts.
turtleblocks.py (GNOME app) is moved alongside the rest of the app code,
and the /usr/bin/turtleblocks launcher now launches it via regular
imports.

The supported use cases will be:

1. Installed as a Sugar activity in the user's home directory.
   As before, all share and lib content is below the activity directory.

2. Installed as a Sugar activity in /usr/share/sugar/activities
   As before, all share and lib content is below the activity directory.

3. Installed as a GNOME application.
   share content owuld normally be found in /usr/share/TurtleBlocks,
   and lib content in /usr/lib/TurtleBlocks.
   This setup is currently broken, before and after this change, but
   this is the first step towards fixing it.

4. From git checkout as GNOME application
   All share and lib content is in the git checkout.