Don't require gconf
turtleblocks.py is already designed to work fine if the GConf client
object is not available, but it still imports it unconditionally.

Adjust so that it works on systems where gconf is not present at all.
Under this configuration, coord_scale had to be set to 0 to avoid the
app running with a very high scale factor.